requirements.txt
pypdf
langchain-community
langchain
langchain-chroma
langchain-huggingface
python-dotenv
tqdm
unstructured
unstructured[pdf]
fastapi
uvicorn
python-multipart
sentence-transformers
chromadb
transformers
tiktoken
threaded
langchain-ollama
llama-index-llms-huggingface
langchain-groq
streamlit